Practice_HTML/CSS_Website
Here is the practice website that I have created over the past week and the weekend to practice HTML and CSS. Please excuse the Hungarian language, but I thought that it wouldn't be a big problem because it is not the text but the functionality that matters. Please note that I had been experimenting with makinf the file "Balatonalmadi.html" responsive with the use of flex boxes. 
Thank you, 
Botond